- added small edits to remove VS compilation warnings

+++ src/Atres.cpp
@@ -128,7 +128,7 @@ namespace Atres
 			if (pos != -1)
 			{
 				f=getFont(name.substr(0,pos));
-				f->setScale(atof(name.substr(pos+1,10).c_str()));
+				f->setScale((float)atof(name.substr(pos+1,10).c_str()));
 			}
 			else throw FontNotFoundException("Font "+name+" does not exist!");
 		}
